rename: gifticon path로 폴더 이름 변경
