+++ app/(challenge)/(다짐 결과)/module/gifticon/useHandleGifticonPage.ts
