+++ app/(challenge)/(다짐 결과)/ui/Gifticon/gifticon.css.ts
